\name{iterateNetwork-package}
\alias{iterateNetwork-package}
\alias{iterateNetwork}
\docType{package}
\title{
Simulation and iteration over network metrics
}
\description{
The function iterateNetwork takes an igraph or network object and calculates network metrics over subsamples of the network.
}
\details{
\tabular{ll}{
Package: \tab iterateNetwork\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2015-10-30\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
Marco T. Bastos
Maintainer: Marco T. Bastos <marco@toledobastos.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
